refactor(sys): move logger impl to `sys/node/logger`
Previously this lived in `src/compiler/sys/logger` but it was only
referenced in `src/sys/node/node-logger.ts`. This moves it to live under
`src/sys/node/logger`, consolidating the 'sys' related stuff in that
directory.
