+++ src/sys/node/index.ts
@@ -1,3 +1,3 @@
-export { createNodeLogger } from './node-logger';
+export { createNodeLogger } from './logger';
 export { setupNodeProcess } from './node-setup-process';
 export { createNodeSys } from './node-sys';
