+++ src/sys/node/logger/index.ts
@@ -1,8 +1,8 @@
+import type { Logger } from '@stencil/core/declarations';
 import fs from 'graceful-fs';
 import path from 'path';
 
-import { createTerminalLogger, TerminalLoggerSys } from '../../compiler/sys/logger/terminal-logger';
-import { Logger } from '../../declarations';
+import { createTerminalLogger, TerminalLoggerSys } from './terminal-logger';
 
 /**
  * Create a logger to run in a Node environment
